Refactore and optimize code of Options Screens
