+++ src/Popup/Router/Screens/Application/OptionsScreen/Components/NotFoundComponent.tsx
@@ -1,13 +1,15 @@
 import React from 'react';
 import classNames from 'classnames';
 
+
 export interface INotFoundOwnProps {
     title?: string;
     description?: string;
     show?: boolean;
 }
 
-export default class NotFoundComponent extends React.Component<INotFoundOwnProps, any> {
+
+export class NotFoundComponent extends React.Component<INotFoundOwnProps, any> {
     render() {
         const {title = '', description = '', show = false} = this.props;
