+++ src/Popup/Router/Screens/Application/OptionsScreen/Components/index.ts
@@ -1,0 +1,2 @@
+export {SearchInputComponent} from './SearchInputComponent';
+export {NotFoundComponent} from './NotFoundComponent';
